TASK: Moved ComponentWithForm class into an own file
